refactor: 분리했던 routes 롤백
