+++ lib/app/router/router.dart
@@ -8,9 +8,6 @@ import 'package:techtalk/presentation/pages/topic_interview/topic_select/intervi
 import 'package:techtalk/presentation/providers/app_user_auth_provider.dart';
 
 part 'router.g.dart';
-part 'routes/main_route.dart';
-part 'routes/sign_in_route.dart';
-part 'routes/sign_up_route.dart';
 
 final rootNavigatorKey = GlobalKey<NavigatorState>();
 
@@ -22,3 +19,68 @@ GoRouter appRouter(WidgetRef ref) => GoRouter(
           : MainRoute.name,
       routes: $appRoutes,
     );
+
+///
+/// Sign In Route
+///
+@TypedGoRoute<SignInRoute>(
+  path: SignInRoute.name,
+  name: SignInRoute.name,
+)
+class SignInRoute extends GoRouteData {
+  const SignInRoute();
+
+  static const String name = '/sign_in';
+  @override
+  Widget build(BuildContext context, GoRouterState state) => const SignInPage();
+}
+
+///
+/// Sign Up Route
+///
+@TypedGoRoute<SignUpRoute>(
+  path: SignUpRoute.path,
+  name: SignUpRoute.path,
+)
+class SignUpRoute extends GoRouteData {
+  const SignUpRoute();
+
+  static const String path = '/sign-up';
+  @override
+  Widget build(BuildContext context, GoRouterState state) {
+    return const SignUpPage();
+  }
+}
+
+///
+/// Main Route
+///
+@TypedGoRoute<MainRoute>(
+  path: MainRoute.name,
+  name: MainRoute.name,
+  routes: [
+    TypedGoRoute<HomeTopicSelectRoute>(
+      path: HomeTopicSelectRoute.name,
+      name: HomeTopicSelectRoute.name,
+    ),
+  ],
+)
+class MainRoute extends GoRouteData {
+  const MainRoute();
+
+  static const String name = '/';
+  @override
+  Widget build(BuildContext context, GoRouterState state) {
+    return MainPage();
+  }
+}
+
+class HomeTopicSelectRoute extends GoRouteData {
+  const HomeTopicSelectRoute();
+
+  static const String name = 'topic-select';
+  @override
+  Widget build(BuildContext context, GoRouterState state) {
+    return InterviewTopicSelectPage();
+  }
+}
